+++ src/main/java/org/halvors/electrometrics/common/block/Block.java
@@ -11,7 +11,7 @@ import org.halvors.electrometrics.Electrometrics;
  *
  * @author halvors
  */
-public class Block extends BlockContainer {
+public abstract class Block extends BlockContainer {
 	protected final String name;
 
 	protected Block(String name, Material material) {
@@ -21,10 +21,5 @@ public class Block extends BlockContainer {
 
 		setUnlocalizedName(name);
 		setCreativeTab(Electrometrics.getTab());
-	}
-
-	@Override
-	public TileEntity createNewTileEntity(World world, int metadata) {
-		return null;
 	}
 }
